Refactored AccoundDataStore to be behind a interface
This is now uses constructor based injection rather than being instanciated. This interface perhaps should be broken up (Interface segregation principal) but here it is fine.

Had issue with VS-code before this relating to updating to .NET 6 - This required the C# Plugin to be switched to newer .net in configuration (    "omnisharp.useModernNet": true) as otherwise intellesense completely broken ~40 mins wasted.